refactor: moved convert to deal modal into separate component
